+++ tests/Stubs/DummyResource.php
@@ -13,9 +13,13 @@ use Illuminate\Http\Resources\Json\JsonResource;
  */
 final class DummyResource extends JsonResource
 {
-    /** @return array<string, int|string> */
+    /** @return array<string, mixed> */
     public function toArray(Request $request): array
     {
+        if (is_array($this->resource)) {
+            return $this->resource;
+        }
+
         return ['id' => $this->id, 'name' => $this->name];
     }
 }